api/requirements.txt
Django==4.2.16
gunicorn==22.0.0
python-dotenv==1.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.11
whitenoise==6.7.0
pytest==8.3.3
pytest-django==4.8.0
model-bakery==1.18.0
